+++ src/commands/set_hours.rs
@@ -5,7 +5,7 @@ use diesel::SqliteConnection;
 use crate::model::hours;
 use crate::model::hours::NewHours;
 
-pub fn execute(when: &NaiveDate, hours_per_week: u8, connection: &mut SqliteConnection) {
+pub fn execute(when: &NaiveDate, hours_per_week: f32, connection: &mut SqliteConnection) {
     let existing = hours::existing(when, connection);
 
     match existing {
@@ -19,13 +19,13 @@ pub fn execute(when: &NaiveDate, hours_per_week: u8, connection: &mut SqliteConn
     );
 }
 
-fn create(when: &NaiveDate, new_hours: u8, connection: &mut SqliteConnection) {
+fn create(when: &NaiveDate, new_hours: f32, connection: &mut SqliteConnection) {
     use crate::schema::hours;
     let date = when.and_hms(0, 0, 0);
 
     let new_hours = NewHours {
         beginning_with: &date,
-        hours_per_week: &(new_hours as i32),
+        hours_per_week: &(new_hours),
     };
 
     diesel::insert_into(hours::table)
@@ -34,11 +34,11 @@ fn create(when: &NaiveDate, new_hours: u8, connection: &mut SqliteConnection) {
         .expect("Unable to set new hours.");
 }
 
-fn update(new_hours: u8, db_id: i32, connection: &mut SqliteConnection) {
+fn update(new_hours: f32, db_id: i32, connection: &mut SqliteConnection) {
     use crate::schema::hours::dsl::*;
 
     diesel::update(hours.filter(id.eq(db_id)))
-        .set(hours_per_week.eq(new_hours as i32))
+        .set(hours_per_week.eq(new_hours))
         .execute(connection)
         .expect("Unable to set new hours.");
 }
